refactor(toolset): move grep helper under tool_invocations
`OutputFilter` and `OutputFilter::apply` are gone with the previous
commit. The only thing that remained in `core/src/toolset/filter.rs`
was the line-level `filter_lines` helper (grep + invert + context),
and its only caller is `apply_fetch_query` (`tool_output_fetch`'s
grep mode). A top-level `mod filter` for a single private helper
used by one consumer is misleading naming.

Inline the helper as `tool_invocations::grep::filter_lines` and
delete the now-empty `filter.rs`. Tests move with the helper.
